+++ buildSrc/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Don't run the workflows when pushing tags to the repository.
+- Set `jvm.variants` and `konan.variants` to `target-only` when importing the
+  project in IntelliJ IDEA.
 
 ## [1.2.0] - 2023-05-16